001: Introduce autoloading and use dotenv
dotenv is used a bit hacky, by being wrapped in another class.
It would be nicer to just use $_ENV, but I don't know how to do this
right now. Would probably either need a overarching execution
to be running or pulling in the dotenv initialization in each file.
This way does basically this, but at least it's centralized.