004: Extract header logic to a new composable
This one is a big one, and once again touches the typing of the
header composables. However, this should be morer or less the last time.
The typing issue on app.vue has been fixed with unref. There are also
now less getters, and a few more reactive types instead.