004: Remove the old app
The 'app' folder is renamed to backend, the PHP files
are moved into an 'api' folder inside that.
The build script is adjusted for that name change, but
the logic is not yet fully changed. The files need to
be built in a correct order and packaged together into a dist folder.